requirements.txt
asgiref==3.9.1
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
Flask==3.1.2
Flask-MySQLdb==2.0.0
fonttools==4.58.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.3
matplotlib==3.10.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
narwhals==1.47.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
pluggy==1.6.0
pygame==2.6.1
PyMySQL==1.1.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.4
